Initial version parser and elements
Initial version of the parser and elements modules. This has been
written without any tests. The test will follow soon.

There is actually one test and some json files that can be used for
testing, but that was written a bit before the parser and most of the
element functionality was written, and only the idea of the builder
pattern existed. First test that will be written is with very small
configs and very small jsons to test the individual bits and pieces.
After that, some complex big configuration can be tested.