Implement all Test and Parsing Logic
Implement the remainder of all the tests for all other elements, and
implemented the logic for arrays. Include a refactor on how the builder
is building the elements, and use an instance of the correct type of
element as soon as possible, even forcing the user to first call a
`ofTypeAbc()` method.

Next commits should add test coverage for the elements themselves and
the element builder.