Automatically ignore $-sign keys in ObjectElements
Automatically ignore any key that starts with a $-sign, as these keys
are normally used for meta data like $comment or $schema. For now, this
does not add much, the same could be achieved with putting
pruneUnknownElements to true, and not add those elements as child
elements of the ObjectElement that contains them. But in the future a
strict mode might be added for validation, where no extra fields are
allowed at all. Then meta-data fields still have their place and should
be ignored and pruned.